feat(firmware): 建立硬體診斷主程式與 PlatformIO 編譯配置
- 建立標準 platformio.ini 並配置 M5StickC-Plus 官方函式庫相依
- 撰寫 src/main.cpp 實作硬體驗證 (LCD 儀表、雙軸搖桿、按鍵、蜂鳴器、RGB 漸變)
- 消除 M5HatMiniJoyC 中 Wire.requestFrom 參數型別歧義警告
- 通過本地 100% 編譯驗證 (RAM 6.8%, Flash 24.5%)

+++ lib/M5HatMiniJoyC/M5HatMiniJoyC.cpp
@@ -123,7 +123,7 @@ uint8_t M5HatMiniJoyC::getI2CAddress(void) {
 
     uint8_t RegValue;
 
-    _wire->requestFrom(_addr, 1);
+    _wire->requestFrom(_addr, (uint8_t)1);
     RegValue = _wire->read();
     return RegValue;
 }
@@ -135,7 +135,7 @@ uint8_t M5HatMiniJoyC::getFirmwareVersion(void) {
 
     uint8_t RegValue;
 
-    _wire->requestFrom(_addr, 1);
+    _wire->requestFrom(_addr, (uint8_t)1);
     RegValue = _wire->read();
     return RegValue;
 }
